+++ common/kernels/imageResize_kernel_st2.cu
@@ -10,7 +10,7 @@
 #define DIAMETER (2*RADIUS+1)
 #define SINGLEDIMINDEX(i,j,width) ((i)*(width) + (j))
 
-__global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* weightedKernel,uint height, uint width,uint resizedRows,uint resizedCols,cudaTextureObject_t tref,int NUM_TEX /*, other arguments */)
+__global__ void resizeKernel_st2(int* outputPixels,int* intermediate, int* weightedKernel,uint height, uint width,uint resizedRows,uint resizedCols,cudaTextureObject_t tref,int NUM_TEX /*, other arguments */)
 {
     // assign id's
     int i = blockIdx.x * blockDim.x + threadIdx.x;
@@ -38,7 +38,7 @@ __global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* w
         j *= yScale;
     }
 
-    float kernelSum = 16.0;
+    int kernelSum = 16.0;
     extern __shared__ int ghb[]; // for per-thread local history
     int my_ghb_index = ((threadIdx.y * blockDim.x) + threadIdx.x) * 3;
     
@@ -48,7 +48,7 @@ __global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* w
         for (int jdx = j; jdx < (j + yScale); jdx++) { // over each element to proc
             if( ((jdx == j + yScale) && ymod == 0) ||
                 ((jdx == j + yScale) && height <= totalY)) break; // same corner case*/
-            float tmp = 0.0;
+            int tmp = 0;
 
             if(idx >=0 && idx < resizedCols
             && jdx >= 0 && jdx < resizedRows-2 ) { // over all row/col
@@ -59,7 +59,7 @@ __global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* w
                     int location = elemToReadFrom + SINGLEDIMINDEX(ii,0,resizedCols);
                     // bounds check #2 for surrounding pix
                     if (location < (width*height) && location >= 0) {
-                        float loaded = intermediate[location];
+                        int loaded = intermediate[location];
                         tmp += loaded * weightedKernel[ii];
                         updateGHB(&(ghb[my_ghb_index]),loaded);
                     }
@@ -69,9 +69,9 @@ __global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* w
                 for(int ii = (DIAMETER-NUM_TEX); ii < DIAMETER;ii++) {
                     int curValueHash = hashGHB(&ghb[my_ghb_index]);
                     int texVal = tex1D<int>(tref,curValueHash);
-                    tmp += (float)(ghb[my_ghb_index+2] + texVal) * weightedKernel[ii];
+                    tmp += (int)(ghb[my_ghb_index+2] + texVal) * weightedKernel[ii];
                 }
-                float avg = tmp / kernelSum;
+                int avg = tmp / kernelSum;
                 outputPixels[elemToWrite] = avg;
             }
         }
