+++ common/kernels/imageResize_kernel_st2.h
@@ -1,2 +1,2 @@
 
-__global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* weightedKernel,uint width, uint height,uint r,uint c,cudaTextureObject_t t,int TEX /*, other arguments */);
+__global__ void resizeKernel_st2(int* outputPixels,int* intermediate, int* weightedKernel,uint width, uint height,uint r,uint c,cudaTextureObject_t t,int TEX /*, other arguments */);
